Fixed most of my checker errors; Fixed correct_test script
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/branches/Release1.3@4366 6adb6e08-d915-0410-941f-83917bcadc18

+++ source/VISUAL/ParamEditor.C
@@ -648,11 +648,11 @@ namespace OpenMS
 				{
 					if(parts[0]!= "")
 					{
-						param_->setMinInt(path,parts[0].toFloat());
+						param_->setMinInt(path,parts[0].toInt());
 					}
 					if(parts[1] != "")
 					{
-						param_->setMaxInt(path,parts[1].toFloat());
+						param_->setMaxInt(path,parts[1].toInt());
 					}
 				}			
 			}
